Extract enum to separate file.
