+++ Core/Core/CourseSync/CourseSyncSettings/Model/CourseSyncFrequency.swift
@@ -1,0 +1,34 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2023-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+enum CourseSyncFrequency: Int, CaseIterable {
+    case daily
+    case weekly
+
+    var stringValue: String {
+        switch self {
+        case .daily: return NSLocalizedString("Daily", comment: "")
+        case .weekly: return NSLocalizedString("Weekly", comment: "")
+        }
+    }
+
+    static var itemPickerData: [ItemPickerSection] {
+        let pickerRows = allCases.map { ItemPickerItem(title: $0.stringValue) }
+        return [ItemPickerSection(items: pickerRows)]
+    }
+}
