+++ Core/CoreTests/CourseSync/CourseSyncSettings/Model/CourseSyncSettingsInteractorTests.swift
@@ -27,31 +27,6 @@ class CourseSyncSettingsInteractorTests: XCTestCase {
         super.tearDown()
     }
 
-    func testSyncFrequencyNames() {
-        XCTAssertEqual(CourseSyncSettingsInteractor.SyncFrequency.weekly.stringValue, "Weekly")
-        XCTAssertEqual(CourseSyncSettingsInteractor.SyncFrequency.daily.stringValue, "Daily")
-    }
-
-    func testSyncFrequenciesToItemPickerData() {
-        let testee = CourseSyncSettingsInteractor.SyncFrequency.itemPickerData
-        XCTAssertEqual(testee.count, 1)
-
-        guard let section = testee.first else {
-            return XCTFail()
-        }
-
-        XCTAssertNil(section.title)
-        XCTAssertEqual(section.items.count, 2)
-        XCTAssertEqual(section.items[0].title, "Daily")
-        XCTAssertEqual(section.items[0].accessibilityIdentifier, nil)
-        XCTAssertEqual(section.items[0].image, nil)
-        XCTAssertEqual(section.items[0].subtitle, nil)
-        XCTAssertEqual(section.items[1].title, "Weekly")
-        XCTAssertEqual(section.items[1].accessibilityIdentifier, nil)
-        XCTAssertEqual(section.items[1].image, nil)
-        XCTAssertEqual(section.items[1].subtitle, nil)
-    }
-
     func testDefaultValues() {
         let testee = CourseSyncSettingsInteractor(storage: defaults)
         XCTAssertFalse(testee.isAutoSyncEnabled.value)
@@ -72,13 +47,13 @@ class CourseSyncSettingsInteractorTests: XCTestCase {
 
         XCTAssertNil(defaults.offlineSyncFrequency)
         testee.syncFrequency.accept(.weekly)
-        XCTAssertEqual(defaults.offlineSyncFrequency, CourseSyncSettingsInteractor.SyncFrequency.weekly.rawValue)
+        XCTAssertEqual(defaults.offlineSyncFrequency, CourseSyncFrequency.weekly.rawValue)
     }
 
     func testReadsValuesFromStorage() {
         defaults.isOfflineAutoSyncEnabled = true
         defaults.isOfflineWifiOnlySyncEnabled = false
-        defaults.offlineSyncFrequency = CourseSyncSettingsInteractor.SyncFrequency.weekly.rawValue
+        defaults.offlineSyncFrequency = CourseSyncFrequency.weekly.rawValue
 
         let testee = CourseSyncSettingsInteractor(storage: defaults)
         XCTAssertTrue(testee.isAutoSyncEnabled.value)
